+++ src/config.c
@@ -7,3 +7,11 @@
 // stall speed
 // pitch
 // trim mappings
+
+#include "config.h"
+
+// control key bindings: CONFIG_KEY_PITCH_UP || CONFIG_MOUSE_BUTTON_PITCH_UP
+bool Config_IsPitchUpCommanded(void)
+{
+    return IsKeyDown(CONFIG_KEY_PITCH_UP) || IsMouseButtonDown(CONFIG_MOUSE_BUTTON_PITCH_UP);
+}