+++ src/gamelogic.c
@@ -1,7 +1,0 @@
-// stall speed calculations
-// heatflux calculations
-// aerodynamic deacceration calculations
-// life calculations
-// control authority mappings? (when stall ==> no/low control authority)
-
-// gamestate (position, velocity, angle, heats, fuel?(for temporary thrust)